Keep only the file-integrity test/demo on this branch
- Delete source-code-integrity MASTG-TEST-0x01 and MASTG-DEMO-0x01
- Rename file-storage-integrity MASTG-TEST-0x02 -> MASTG-TEST-0x01 and
  MASTG-DEMO-0x02 -> MASTG-DEMO-0x01 (no other test/demo uses 0x01)
- Update MASTG-TEST-0090 covered_by accordingly
