+++ demos/ios/MASVS-RESILIENCE/MASTG-DEMO-0x01/MastgTest.swift
@@ -1,55 +1,74 @@
-// SUMMARY: This sample makes a security-sensitive decision (a license check) but never verifies
-// the integrity of its own code at runtime. An attacker can patch the compiled check on a
-// jailbroken device and the app will not detect the modification.
+// SUMMARY: This sample stores sensitive data in a file in the app's Documents directory and
+// later reads it back without computing or verifying any integrity value (HMAC or signature).
+// An attacker who modifies the file on a jailbroken device can tamper with the data undetected.
 //
-// For contrast, the file also contains a PASS routine that hashes the app's own __TEXT/__text
-// section at runtime. Keeping both cases in one binary makes the difference visible in the
-// disassembly: the PASS routine references dladdr/getsectiondata/CC_SHA256, the FAIL routine
-// references none of them.
+// For contrast, the file also contains a PASS routine that stores the same data with an appended
+// HMAC-SHA256 and verifies it on read. Keeping both cases in one binary makes the difference
+// visible in the disassembly: the PASS routine references CryptoKit's HMAC, the FAIL routine
+// references no integrity API at all.
 
 import Foundation
-import CommonCrypto
-import MachO
+import CryptoKit
 
 struct MastgTest {
-    // FAIL case: a security-sensitive gate. An attacker who patches the compiled comparison can
-    // flip the result; nothing verifies the integrity of this code at runtime.
-    static func isLicenseValid(_ licenseKey: String) -> Bool {
-        return licenseKey == "MAS-PREMIUM-2025"
-    }
+    // PASS case: store data with an appended HMAC-SHA256 tag and verify it before trusting the
+    // data on read. A mismatch (for example after the file is tampered with) is detected. In a
+    // real app the key would be held in the Keychain rather than generated in-process.
+    static func storeWithIntegrity(_ data: Data, at fileURL: URL, using key: SymmetricKey) -> String {
+        let tagLength = 32 // HMAC-SHA256 tag size in bytes
+        do {
+            let mac = HMAC<SHA256>.authenticationCode(for: data, using: key)
+            try (data + Data(mac)).write(to: fileURL)
 
-    // PASS case: hash the app's own __TEXT/__text section at runtime. A real implementation would
-    // compare this digest against a securely stored reference value and react if they differ,
-    // letting the app detect binary patching. This routine references dladdr, getsectiondata, and
-    // CC_SHA256, so unlike isLicenseValid these integrity APIs are visible in the binary.
-    static func currentCodeHash() -> String? {
-        var info = Dl_info()
-        guard dladdr(#dsohandle, &info) != 0, let base = info.dli_fbase else { return nil }
-        let header = base.assumingMemoryBound(to: mach_header_64.self)
-        var size: UInt = 0
-        guard let text = getsectiondata(header, "__TEXT", "__text", &size) else { return nil }
-        var digest = [UInt8](repeating: 0, count: Int(CC_SHA256_DIGEST_LENGTH))
-        CC_SHA256(text, CC_LONG(size), &digest)
-        return digest.map { String(format: "%02x", $0) }.joined()
+            let stored = try Data(contentsOf: fileURL)
+            let payload = stored.prefix(stored.count - tagLength)
+            let tag = stored.suffix(tagLength)
+            let valid = HMAC<SHA256>.isValidAuthenticationCode(Data(tag), authenticating: Data(payload), using: key)
+            return valid ? "verified" : "tampering detected"
+        } catch {
+            return "failed: \(error.localizedDescription)"
+        }
     }
 
     static func mastgTest(completion: @escaping (String) -> Void) {
-        // FAIL: [MASTG-TEST-0x01] The app makes a security-sensitive decision without computing
-        // or verifying a hash over its own __TEXT/__text section, so binary patching goes undetected.
-        let providedKey = "INVALID-KEY"
-        let unlocked = isLicenseValid(providedKey)
+        // FAIL: [MASTG-TEST-0x01] The app writes sensitive data to disk and later reads it back
+        // without computing or verifying an HMAC or signature, so it cannot detect tampering.
 
-        // PASS: compute an integrity hash over the app's own code for contrast. The FAIL decision
-        // above produces no integrity evidence; this routine returns a digest that a real check
-        // would compare against a trusted reference.
-        let codeHash = currentCodeHash() ?? "unavailable"
+        let fileManager = FileManager.default
+        guard let documents = fileManager.urls(for: .documentDirectory, in: .userDomainMask).first else {
+            completion("Could not locate the Documents directory")
+            return
+        }
+        let fileURL = documents.appendingPathComponent("user_profile.json")
+
+        // Store sensitive data without any integrity protection
+        let sensitiveData = #"{"username":"alice","role":"user","premium":false}"#.data(using: .utf8)!
+
+        do {
+            try sensitiveData.write(to: fileURL)
+        } catch {
+            completion("Failed to write file: \(error.localizedDescription)")
+            return
+        }
+
+        // Later, the app reads the data back and trusts it without verifying its integrity
+        guard let loaded = try? Data(contentsOf: fileURL),
+              let contents = String(data: loaded, encoding: .utf8) else {
+            completion("Failed to read the file back")
+            return
+        }
+
+        // PASS: for contrast, store the same data with an HMAC and verify it on read. The FAIL path
+        // above produces no integrity value; this routine returns the verification result.
+        let protectedURL = documents.appendingPathComponent("user_profile_protected.json")
+        let integrityResult = storeWithIntegrity(sensitiveData, at: protectedURL, using: SymmetricKey(size: .bits256))
 
         let value = """
-        Provided key   : \(providedKey)
-        Premium access : \(unlocked ? "GRANTED" : "DENIED")
+        Stored file : \(fileURL.path)
+        Contents    : \(contents)
 
-        FAIL (no integrity check) : the decision above was made without verifying the app's code.
-        PASS (code integrity hash): \(codeHash)
+        FAIL (no integrity check) : the data above was read back without verifying any HMAC or signature.
+        PASS (HMAC verified)      : \(integrityResult)
         """
         completion(value)
     }
