+++ demos/ios/MASVS-RESILIENCE/MASTG-DEMO-0x01/run.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-r2 -q -i source_integrity.r2 -A MASTestApp > output.asm
+r2 -q -i storage_integrity.r2 -A MASTestApp > output.asm
