--- demos/ios/MASVS-RESILIENCE/MASTG-DEMO-0x01/source_integrity.r2
@@ -1,21 +1,0 @@
-e scr.color=0
-e scr.interactive=false
-e bin.relocs.apply=true
-
-?e
-?e Evidence of a security-sensitive routine (license key string literal):
-izz~MAS-PREMIUM
-
-?e
-?e Searching the import table for source code integrity APIs:
-
-?e
-?e Binary base address resolution (dladdr):
-ii~dladdr
-
-?e
-?e Cryptographic hash functions over code (CC_MD5 / CC_SHA256 / CC_SHA512):
-ii~CC_
-
-?e
-?e (Any integrity APIs listed above are referenced only by an unrelated routine and are not used to verify the license check.)
