+++ demos/ios/MASVS-RESILIENCE/MASTG-DEMO-0x01/storage_integrity.r2
