--- demos/ios/MASVS-RESILIENCE/MASTG-DEMO-0x02/output.asm
@@ -1,13 +1,0 @@
-
-Evidence that the app stores data on disk (filename string literal):
-103 0x00009d80 0x100009d80 17  18   3.__TEXT.__cstring         ascii   user_profile.json
-
-Searching the import table for file storage integrity APIs:
-
-HMAC (CCHmac / CCHmacFinal):
-
-Hash functions (CC_SHA256 / CC_SHA512):
-
-Asymmetric signing (SecKeyCreateSignature):
-
-(No matches under the integrity APIs means none are referenced by the app.)
