--- demos/ios/MASVS-RESILIENCE/MASTG-DEMO-0x02/run.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-r2 -q -i storage_integrity.r2 -A MASTestApp > output.asm
