+++ tests-beta/ios/MASVS-RESILIENCE/MASTG-TEST-0x01.md
@@ -1,9 +1,10 @@
 ---
 platform: ios
-title: References to Source Code Integrity Check APIs
+title: References to File Storage Integrity Check APIs
 id: MASTG-TEST-0x01
+apis: [CCHmac, CCHmacFinal, CC_SHA256, CC_SHA512, SecKeyCreateSignature]
 type: [static, code, manual]
-weakness: MASWE-0104
+weakness: MASWE-0105
 false_negative_prone: true
 profiles: [R]
 knowledge: [MASTG-KNOW-0086]
@@ -12,20 +13,18 @@ best-practices: [MASTG-BEST-0x01]
 
 ## Overview
 
-iOS apps can implement runtime source code integrity checks to detect if the binary has been tampered with. These checks typically parse the Mach-O binary structure to locate the `__TEXT/__text` section, compute a hash over it, and compare that hash against a reference value (see @MASTG-KNOW-0086). If the app does not implement such checks, an attacker who patches the binary can go undetected.
+iOS apps can protect the integrity of data they store on the device (e.g., files in the Documents directory, `UserDefaults`/`NSUserDefaults`, or databases) by computing an HMAC or a digital signature over the data and verifying it before use (see @MASTG-KNOW-0086). If the app does not implement such checks, an attacker who modifies stored data can go undetected.
 
-This test verifies that the app references APIs commonly used to implement source code integrity checks, such as `dladdr` for resolving the binary base address, Mach-O header parsing structures (`mach_header`, `load_command`), and cryptographic hash functions applied to code sections (e.g., `CC_SHA256`).
-
-Note that [Apple's code signing](https://developer.apple.com/documentation/xcode/using-the-latest-code-signature-format) and DRM (FairPlay) provide some level of integrity protection at the OS level, but additional in-app runtime checks raise the bar for attackers who operate on jailbroken devices where these protections may be bypassed.
+This test verifies that the app references APIs commonly used to implement file storage integrity checks, such as `CCHmac`, `CCHmacFinal` (HMAC via CommonCrypto), `CC_SHA256`, `CC_SHA512`, or [`SecKeyCreateSignature`](https://developer.apple.com/documentation/security/seckeycreatesignature(_:_:_:_:)) (asymmetric signing).
 
 **Example Attack Scenario:**
 
-Suppose a banking app relies only on OS-level code signing and implements no runtime integrity check.
+Suppose an app stores a usage counter or entitlement flag in `UserDefaults` and trusts it without verifying its integrity.
 
-1. An attacker reverse engineers the app and patches the binary to disable a security control (for example, jailbreak detection) using @MASTG-TECH-0090.
-2. The attacker re-signs the modified app with their own certificate using @MASTG-TECH-0092.
-3. The attacker installs the patched app on a jailbroken device, where OS-level code signing no longer protects it.
-4. Because the app never verifies the integrity of its own code at runtime, the tampering goes undetected and the attacker can run the app with the security control removed.
+1. An attacker uses @MASTG-TECH-0059 to locate the app's data directories on a jailbroken device.
+2. The attacker modifies the stored value (for example, resets a trial counter or flips a "premium" flag).
+3. Because the app never verifies an HMAC or signature over the stored data, it loads the tampered value as authentic.
+4. The attacker bypasses the intended restriction, gaining access to functionality or content they should not have.
 
 ## Steps
 
@@ -34,19 +33,19 @@ Suppose a banking app relies only on OS-level code signing and implements no run
 
 ## Observation
 
-The output should include any references to source code integrity check APIs such as `dladdr`, `mach_header`, `LC_SEGMENT`, and cryptographic hash functions (e.g., `CC_MD5`, `CC_SHA256`, `CC_SHA512`) applied to code sections.
+The output should include a list of locations where the relevant APIs are used.
 
 ## Evaluation
 
-The test case fails if the app contains no references to source code integrity check APIs.
+The test case fails if the app stores sensitive data locally (in files, `UserDefaults`/`NSUserDefaults`, or databases) in use cases that are not covered by a corresponding file storage integrity check.
 
 **Further Validation Required:**
 
-Inspect each reported code location using @MASTG-TECH-0076 to determine whether the referenced APIs actually implement a source code integrity check:
+These APIs are commonly linked for unrelated purposes (for example, networking, analytics, or generic checksums), so their mere presence does not confirm a storage integrity mechanism. Inspect each reported code location using @MASTG-TECH-0076 to determine whether the referenced APIs actually protect stored data:
 
-- Determine whether the hash is computed over the `__TEXT/__text` section (rather than, for example, hashing unrelated data for caching or networking).
-- Determine whether the computed hash is compared against a stored reference value and the app reacts when the comparison fails.
+- Determine whether the HMAC or signature is computed over data the app reads back from local storage (files, `UserDefaults`/`NSUserDefaults`, or a database).
+- Determine whether the app verifies the HMAC or signature before using the data and reacts when verification fails.
 
 **Expected False Negatives:**
 
-This test may produce false negatives if the integrity check is built using patterns not covered by the analysis (for example, a custom hash function instead of `CC_SHA256`). In such cases, the absence of findings does not guarantee the absence of an integrity check, and additional manual reverse engineering may be required.
+This test may produce false negatives if the integrity check relies on a third-party library, a custom implementation, or APIs not covered by the analysis. In such cases, the absence of findings does not guarantee the absence of a storage integrity check, and additional manual reverse engineering may be required.
