+++ tests/ios/MASVS-RESILIENCE/MASTG-TEST-0090.md
@@ -10,7 +10,7 @@ masvs_v1_levels:
 - R
 profiles: [R]
 status: deprecated
-covered_by: [MASTG-TEST-0x01, MASTG-TEST-0x02]
+covered_by: [MASTG-TEST-0x01]
 deprecation_note: New version available in MASTG V2
 ---
